requirements.txt
asgiref==3.4.1
botocore==1.21.53
certifi==2021.5.30
charset-normalizer==2.0.6
django-js-asset==1.2.2
django-mptt==0.13.4
easy-thumbnails==2.7.1
gunicorn==20.1.0
idna==3.2
jmespath==0.10.0
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.3
recommend==0.2.2
requests==2.26.0
s3transfer==0.5.0
scipy==1.7.1
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
ipython==7.28.0
channels==3.0.4
psycopg2-binary==2.9.1
django-widget-tweaks
djangorestframework
django-cors-headers
django-filter
django-filer
django-hitcount
